Moved DebugLoggingUserModule into UserModules subdirectory
