+++ src/CMakeLists.txt
@@ -1,5 +1,6 @@
 
 file(GLOB source_files *.cc)
-add_library(app-framework ${source_files})
+add_library(app-framework SHARED ${source_files} )
 
 add_subdirectory(Buffers)
+add_subdirectory(UserModules)
