+++ src/UserModules/CMakeLists.txt
@@ -1,0 +1,3 @@
+
+file(GLOB source_files *.cc)
+add_library(app-framework_UserModules SHARED ${source_files} )
