+++ unittest/DAQProcess_test.cc
@@ -15,7 +15,7 @@ std::unique_ptr<CommandFacility> CommandFacility::handle_ = std::unique_ptr<Comm
 
 #define BOOST_TEST_MODULE DAQProcess_test
 
-#include <boost/test/auto_unit_test.hpp>
+#include <boost/test/unit_test.hpp>
 
 BOOST_AUTO_TEST_SUITE(DAQProcess_test)
